Breakout components and update babel and eslinting
