+++ babel.config.js
@@ -8,5 +8,6 @@ module.exports = {
         },
       },
     ],
+    '@babel/preset-react',
   ],
 };
